Moved leaderboard routes
