+++ src/server/src/leaderboard/router.ts
@@ -1,0 +1,7 @@
+import { Router } from 'express';
+import { controller } from 'src/leaderboard/controller';
+
+export const router = Router();
+
+router.get('/leaderboard/item', controller.getItemLeaderboard);
+router.get('/leaderboard/ability', controller.getAbilityLeaderboard);
