+++ src/server/src/router.ts
@@ -1,6 +1,7 @@
 import { Router } from 'express';
 import { controller } from 'src/controller';
 import { router as playerRouter } from 'src/player/router';
+import { router as leaderboardRouter } from 'src/leaderboard/router';
 
 export const router = Router();
 
@@ -9,13 +10,11 @@ router.get('/status', (req, res, next) => {
 });
 
 router.use(playerRouter);
+router.use(leaderboardRouter);
 
 router.get('/item', controller.getItemMetadata);
 
 router.get('/ability', controller.getAllAbilityMetadata);
-
-router.get('/leaderboard/item', controller.getItemLeaderboard);
-router.get('/leaderboard/ability', controller.getAbilityLeaderboard);
 
 router.get('/search/player', controller.searchPlayer);
